Katarina

See merge request !11

+++ src/state.h
@@ -15,11 +15,7 @@
 
 struct stateOfMachine {
     uint32_t registers[17];
-    uint32_t *mem;
+    uint8_t *mem;
 };
-
-
-
-
 
 #endif //C_GROUP_31_STATE_H
